Adds unit test for rcp multithreading to address debug node tracing.
I changed the addNewRCPNode and removeRCPNode static functions to be locked by a mutex to prevent thread conflicts.
I moved the node_->delete_obj() call to be inside removeRCPNode and protected by the mutex for debug only.
This seems to address the problems with thread safety of debug node tracing.