created shared rcp test and a std::shared_ptr test which mirrors it exactly.
This is to address unit test 3 that Ross suggested.
The singleton shares an rcp to threads.
If no thread uses the object it will allow the memory to be deallocated.

At this stage the std::shared_ptr case works.
The rcp case has a rare conflict which I will be tracking down next.