RCP Unit Tests
This has the initial set of 7 tests we proposed trying.
Fixes are in place for the 1st 5.
The 6th is implemented and works but there are further ways we could break this - we may not need to pending discussion.
The 7th shows how a weak dangling ptr can cause thread issues. Working on ideas to resolve this.
This also has some possibly temporary code to consider tracking strong and total instead of strong and weak which might make a more elegant pipeline.