Fixed the debug node tracing mutex for statics.
This fix repairs the mutex to properly handle RCP statics.
The issue is the order of deallocation for the static mutex and a static RCP could be wrong, causing the mutex to throw an exception/crash if deallocated.

Now the mutex is allocated with a new/delete pattern - follows the rcp_node_list() method.