Fix Array class to have thread safe dangling reference detection for iterators.
Concurrent modification and iteration of array will now detect
dangling references in a thread safe manner.

This does not provide protection against undefined behaviors for concurrent
reading of an array while it is being modified.
For example, array->at[0] can lead to undefined access results.
That issue will be worked on in future updates.

Note this commit is planned to be the last old master commit and this branch will be moved to develop now.