Amesos2: Implement Umfpack wrappers for Amesos2
The wrappers are modeled on the SuperLu wrappers.
The tests also closely follow the SuperLu format.

One important diffeence is Umfpack supports only
double types, not float types. Conversions have been
implemented and a warning is given in CMake if
float is on and double is off.

Status:
Need to resolve warnings for AMD due to multiple defines.
Determine if we should be handling float on/double off case.
Determine full testing procedure.