Zoltan2: Initial directory setup
The directory handles std::vectors of a type
and can have variable length. An add mode is
also implemented.

The current unit test creates std::vector<int>
of variable length for the user_t, fills the
elements with count 1, then calls add which sums
the arrays to provide a count of total processes
sharing the same gid. The test validates the
find call obtains the correct value by checking
one of the array elements (all currently the same).

reverse execution works for variable length
without prior knowledge of the sizes. They
are collected from the process which stores
the data for that node.

The other 3 tests (Original,Relic,Tpetra) will
pass when the test is reconfigured for non-array
mode without using resize. These will be developed
and used for performance comparisons and eventually
deleted.

The remove feature was previously working but
is temporarily turned off  - will be restored for
future commits.

There are a several places which calculate offsets in a slow
manual way to be refactored or stored in arrays later,
but that will be inefficient in the present form.

There are also many places which need further cleaning up
to make the style more C++ like. Especially the handling
of ptrs for the message types.