KokkosKernels: Working prototype of YAML archiver
This implements a working prototype of the new YAML
performance archiver in KokkosKernels.

The blas1 performance test was duplicated for demo
purposes and modifications were made so that it would
run using the archiver.

Currently the archiver extracts test times
from the already in place Teuchos timers though
presumably all the Teuchos references will be going
away later.