Zoltan2: Directory Updates
Fixes a bunch of type mismatches I wasn't properly checking for earlier.
Fixes a few errors for the various modes and removes some unneeded code.
Upgrades the tests to both an Add and Replace check.
Started setup for an Aggregate mode as well but not implemented.
Adds some debugging tools to the communicator class.

Also added more careful comments for execute_wait.
Next step will be delve into this more carefully - I have this working
for variable sizes reverse execution but possibly this can be
done in a simpler way. The relevant code is all in execute_wait and
once I understand it better I expect it can be adapted to the current
code. I am not sure if the original directory is already set up to
handle this case or if we need some new code like this to make it work.

There is a bug for small total ID count (such as 10) which is caused by
no_send_buff becoming 1 in constructor for some cases due to the gid
list being simple and happening to have all gids sorted by process.
The bug hits in execute_wait so I am fairly certain this is all related
to the same issues.

Other tasks to also be done:
  Implement Aggregate
  Check/implement gid of multiple components - such as int[4].
  Check edges cases more carefully (for example set to rec 0 items)
  Setup a PR

So this is a checkpoint and work in progress.

Note all 4 modes should pass unit tests (Kokkos, Original, Relic, Tpetra) but
only Kokkos is currently set up to be using the new variable array length
mode. I may fix Relic to do as well - Tpetra and Original won't ever support
that anyways.