Amesos2: Tacho ideas for optimization
This implemented an optimized view when types match,
a static_cast loop to copy when they don't.

A few things I'd like to discuss further on this,
such as the use of a forced cast to make the line of
code for direct view compile when the optimization
is not directly possible. I'm not sure if there is a
better way to handle this. Questions are noted in
line 113 of Amesos2_Tacho_def.hpp.