Zoltan2: Multijagged restore OpenMP checkpoint
Now OpenMP off and on both work.

This is still thread 1 so some disabled code is apparetly not relevant.

Added a macro RESTORE_NO_OMP_CODE which flips back on some of the
original code to preserve working behavior.

However this checkpoint does convert a bunch of gid
code to Kokkos::Views so a step forward.