Zoltan2: Remove all the HAVE_ZOLTAN2_OMP calls
All tests pass for par/ser with OpenMP or Serial
but now run through the same code path.

Exception is the mj_driver_test which has
intermittent failures due to being out of the
imbalance range. Wasn't clear yet if there is
an error or the thread allocations work in
such a way to change these. Note that develop
branch will also demonstrate varied results
for the imbalance. Needs to be investigated.