Zoltan2: Fix some issues with the lambdas.
This can be further improved but gets the tests to passing
state at least. Some code needs to be shoved around to prevent
lambdas in constructors so probably some inefficiencies have
been introduced with this change.

Note at this point I'm still getting intermittent failures
for Zoltan2_mjDriverTest_MPI_3 but that happens with OpenMP
off as well. So it's a bit surprising that is happening but
will resolve.