Zoltan2: Cuda check point
Remove the name() call on the custom node.
With ETI on that won't exist - luckily it seems I can otherwise use node.
May eventually change this to just pass device.

This will pass for cuda with UVM on and fail some tests with it off.
This is what we expected - now I can investigate the errors in the
code for UVM off.

Was expecting more tests to fail for UVM off - but will see.