Zoltan2: Add backwards compatibility for adapters to use kokkos.
If an adapter does not have kokkos view capability, the kokkos
views will be generated using the non-kokkos calls.

Now it also tests an adapter with only kokkos calls.

Zoltan2_VectorAdapter now has backwards and forward compatibility
in same pattern as Zoltan2_Adapter.

rcb is also run as a control along with multijagged.