KokkosGraph_Distance1Color.hpp: changes of the form,
-  typedef typename Impl::GraphColor
+  typedef typename ::KokkosGraph::Impl::GraphColor

since GraphColor is not in KokkosGraph::Experimental.

Mark endpoints of namespace Experimental in files of src/graph to help
understanding where the namespaces end.

This commit is part of kokkos-kernels, but a pull request has not been
generated because upstream has changed too much.  A diff against what
is in trilinos is attached as graph.txt to kokkos-kernels issue #978.
Will try again when trilinos gets the kokkos-kernels updates.